+++ CHANGELOG.md
@@ -26,14 +26,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- New `--logs.http-traffic` option
 - Support direct local connection to JDownloader using their deprecated API
 - `wreq` as optional HTTP backend for browser impersonation (via the `wreq` extra)
 
 ### Changed
 
 - Log HTTP requests made to JDownloader
+- Logs of HTTP requests/responses will be emitted at INFO instead of DEBUG level
 - If scraping has already started, sending `SIGINT` (`Ctrl + C`) to the process will stop all scraping but will continue with post-runtime actions
   (sorting, hashing, send notifications, etc..). Press `Ctrl + C` again to force quit
+- Reduced the maximum size the log file can have to be sent as attachment in notifications, from 25 MB to 20 MB.
 
 ## [10.5.0] - 2026-08-15